ASDA

+++ train_gpt_02_convolve_embed.py
@@ -374,7 +374,8 @@ class ConvEmbedding(nn.Module):
         nn.init.normal_(self.token_embedding.weight, mean=0, std=self.embedding_dim ** -0.5)
         nn.init.xavier_uniform_(self.conv.weight)
         nn.init.zeros_(self.conv.bias)
-        # Convert bias to bfloat16 to match input type
+        # Convert both weights and bias to bfloat16 to match input type
+        self.conv.weight.data = self.conv.weight.data.to(torch.bfloat16)
         self.conv.bias.data = self.conv.bias.data.to(torch.bfloat16)
 
     def forward(self, x):
